@@ -1,8 +1,6 @@
-import * as fs from 'fs';
 import autobind from 'autobind-decorator';
 import Module from '../../module';
 import serifs from '../../serifs';
-import * as tmp from 'tmp';
 import { genMaze } from './gen-maze';
 import { renderMaze } from './render-maze';
 import Message from '../../message';
@@ -41,26 +39,18 @@ export default class extends Module {
 	}
 
 	@autobind
-	private createTemp(): Promise<[string, any]> {
-		return new Promise<[string, any]>((res, rej) => {
-			tmp.file((e, path, fd, cleanup) => {
-				if (e) return rej(e);
-				res([path, cleanup]);
-			});
-		});
-	}
-
-	@autobind
 	private async genMazeFile(seed, size?): Promise<any> {
 		this.log('Maze generating...');
 		const maze = genMaze(seed, size);
 
 		this.log('Maze rendering...');
-		const [temp] = await this.createTemp();
-		await renderMaze(seed, maze, fs.createWriteStream(temp));
+		const data = renderMaze(seed, maze);
 
 		this.log('Image uploading...');
-		const file = await this.ai.upload(fs.createReadStream(temp));
+		const file = await this.ai.upload(data, {
+			filename: 'maze.png',
+			contentType: 'image/png'
+		});
 
 		return file;
 	}
